pid_monitor/tests/requirements.txt
matplotlib==2.1.1
psutil==5.4.2
pyserial==3.4